Cruft/style and include the global msg name in parsed output
After an encode or decode the Parser.messages[] objects will have a
`name` attr with the global message name.  It was not visible before so
you didn't actually know what the message was related to, but also
made it impossible to rebuild a FIT file using just the message descs.